remove couts, fix increment input parameter

+++ src/lib/storage/chunk.cpp
@@ -208,7 +208,7 @@ void Chunk::set_statistics(const std::shared_ptr<ChunkStatistics>& chunk_statist
   _statistics = chunk_statistics;
 }
 
-void Chunk::increase_invalid_row_count(const uint32_t count) {
+void Chunk::increase_invalid_row_count(const uint64_t count) {
   _invalid_row_count += count;
 }
